delete unit test and intergration test created files
