+++ broker/src/test/java/org/apache/rocketmq/broker/BrokerControllerTest.java
@@ -18,10 +18,14 @@
 package org.apache.rocketmq.broker;
 
 import org.apache.rocketmq.common.BrokerConfig;
+import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.remoting.netty.NettyClientConfig;
 import org.apache.rocketmq.remoting.netty.NettyServerConfig;
 import org.apache.rocketmq.store.config.MessageStoreConfig;
+import org.junit.After;
 import org.junit.Test;
+
+import java.io.File;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -43,5 +47,10 @@ public class BrokerControllerTest {
             brokerController.start();
             brokerController.shutdown();
         }
+    }
+
+    @After
+    public void destory(){
+        UtilAll.deleteFile(new File(new MessageStoreConfig().getStorePathRootDir()));
     }
 }
