+++ broker/src/test/java/org/apache/rocketmq/broker/filter/ConsumerFilterManagerTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.rocketmq.broker.filter;
 
+import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.filter.ExpressionType;
 import org.apache.rocketmq.common.filter.FilterAPI;
 import org.apache.rocketmq.common.protocol.heartbeat.SubscriptionData;
@@ -232,7 +233,7 @@ public class ConsumerFilterManagerTest {
             assertThat(filterData.isDead()).isTrue();
             assertThat(filterData.getCompiledExpression()).isNotNull();
         } finally {
-            deleteDirectory("./unit_test");
+            UtilAll.deleteFile(new File("./unit_test"));
         }
     }
 
@@ -269,23 +270,8 @@ public class ConsumerFilterManagerTest {
 
             assertThat(topicData).isNullOrEmpty();
         } finally {
-            deleteDirectory("./unit_test");
+            UtilAll.deleteFile(new File("./unit_test"));
         }
     }
 
-    protected void deleteDirectory(String rootPath) {
-        File file = new File(rootPath);
-        deleteFile(file);
-    }
-
-    protected void deleteFile(File file) {
-        File[] subFiles = file.listFiles();
-        if (subFiles != null) {
-            for (File sub : subFiles) {
-                deleteFile(sub);
-            }
-        }
-
-        file.delete();
-    }
 }
