+++ broker/src/test/java/org/apache/rocketmq/broker/filter/MessageStoreWithFilterTest.java
@@ -18,6 +18,7 @@
 package org.apache.rocketmq.broker.filter;
 
 import org.apache.rocketmq.common.BrokerConfig;
+import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.filter.ExpressionType;
 import org.apache.rocketmq.common.message.MessageDecoder;
 import org.apache.rocketmq.common.message.MessageExt;
@@ -174,22 +175,6 @@ public class MessageStoreWithFilterTest {
         return msgs;
     }
 
-    protected void deleteDirectory(String rootPath) {
-        File file = new File(rootPath);
-        deleteFile(file);
-    }
-
-    protected void deleteFile(File file) {
-        File[] subFiles = file.listFiles();
-        if (subFiles != null) {
-            for (File sub : subFiles) {
-                deleteFile(sub);
-            }
-        }
-
-        file.delete();
-    }
-
     protected List<MessageExtBrokerInner> filtered(List<MessageExtBrokerInner> msgs, ConsumerFilterData filterData) {
         List<MessageExtBrokerInner> filteredMsgs = new ArrayList<MessageExtBrokerInner>();
 
@@ -301,7 +286,7 @@ public class MessageStoreWithFilterTest {
         } finally {
             master.shutdown();
             master.destroy();
-            deleteDirectory(storePath);
+            UtilAll.deleteFile(new File(storePath));
         }
     }
 
@@ -386,7 +371,7 @@ public class MessageStoreWithFilterTest {
         } finally {
             master.shutdown();
             master.destroy();
-            deleteDirectory(storePath);
+            UtilAll.deleteFile(new File(storePath));
         }
     }
 }
