+++ common/src/main/java/org/apache/rocketmq/common/UtilAll.java
@@ -495,4 +495,19 @@ public class UtilAll {
             throw new RuntimeException("Can not get local ip", e);
         }
     }
+
+    public static void deleteFile(File file) {
+        if (!file.exists()) {
+            return;
+        }
+        if (file.isFile()) {
+            file.delete();
+        } else if (file.isDirectory()) {
+            File[] files = file.listFiles();
+            for (File file1 : files) {
+                deleteFile(file1);
+            }
+            file.delete();
+        }
+    }
 }
