+++ store/src/test/java/org/apache/rocketmq/store/ConsumeQueueTest.java
@@ -18,6 +18,7 @@
 package org.apache.rocketmq.store;
 
 import org.apache.rocketmq.common.BrokerConfig;
+import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.message.MessageDecoder;
 import org.apache.rocketmq.store.config.MessageStoreConfig;
 import org.apache.rocketmq.store.stats.BrokerStatsManager;
@@ -131,22 +132,6 @@ public class ConsumeQueueTest {
         }
     }
 
-    protected void deleteDirectory(String rootPath) {
-        File file = new File(rootPath);
-        deleteFile(file);
-    }
-
-    protected void deleteFile(File file) {
-        File[] subFiles = file.listFiles();
-        if (subFiles != null) {
-            for (File sub : subFiles) {
-                deleteFile(sub);
-            }
-        }
-
-        file.delete();
-    }
-
     @Test
     public void testConsumeQueueWithExtendData() {
         DefaultMessageStore master = null;
@@ -220,7 +205,7 @@ public class ConsumeQueueTest {
         } finally {
             master.shutdown();
             master.destroy();
-            deleteDirectory(storePath);
+            UtilAll.deleteFile(new File(storePath));
         }
     }
 }
