+++ store/src/test/java/org/apache/rocketmq/store/index/IndexFileTest.java
@@ -20,8 +20,11 @@
  */
 package org.apache.rocketmq.store.index;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
+
+import org.apache.rocketmq.common.UtilAll;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -42,6 +45,8 @@ public class IndexFileTest {
         boolean putResult = indexFile.putKey(Long.toString(400), 400, System.currentTimeMillis());
         assertThat(putResult).isFalse();
         indexFile.destroy(0);
+        File file = new File("100");
+        UtilAll.deleteFile(file);
     }
 
     @Test
@@ -62,5 +67,7 @@ public class IndexFileTest {
         assertThat(phyOffsets).isNotEmpty();
         assertThat(phyOffsets.size()).isEqualTo(1);
         indexFile.destroy(0);
+        File file = new File("200");
+        UtilAll.deleteFile(file);
     }
 }
